Backend/flask-server/requirements.txt
Flask==2.0.1
Flask-PyMongo==2.3.0
pandas==1.4.0
flask-cors==3.1.1
PyJWT==2.3.0
